+++ lib/atom_utils.dart
@@ -25,6 +25,8 @@ final JsObject _fs = require('fs');
 /// 'darwin', 'freebsd', 'linux', 'sunos' or 'win32'
 final String platform = _process['platform'];
 
+String get chromeVersion => _process['versions']['chrome'];
+
 final bool isWindows = platform.startsWith('win');
 final bool isMac = platform == 'darwin';
 final bool isLinux = !isWindows && !isMac;